competitions/examples/requirements.txt
fess38.data_processing @ file:///usr/local/lib/kaggle/data_processing
